agents/market_researcher/requirements.txt
google-cloud-aiplatform[adk,agent_engines]>=1.90.0
google-adk
python-dateutil
a2a_common-0.1.0-py3-none-any.whl